refactor: menu 패키지 정리
