+++ src/test/java/kitchenpos/menu/ui/MenuRestControllerTest.java
@@ -1,9 +1,11 @@
-package kitchenpos.ui;
+package kitchenpos.menu.ui;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import kitchenpos.UTF8MockMvcTest;
-import kitchenpos.application.MenuGroupService;
-import kitchenpos.domain.MenuGroup;
+import kitchenpos.menu.application.MenuService;
+import kitchenpos.menu.domain.Menu;
+import kitchenpos.menu.domain.MenuProduct;
+import kitchenpos.menu.ui.MenuRestController;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -11,6 +13,7 @@ import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.http.MediaType;
 import org.springframework.test.web.servlet.MockMvc;
 
+import java.math.BigDecimal;
 import java.util.Arrays;
 
 import static org.hamcrest.Matchers.containsString;
@@ -21,8 +24,8 @@ import static org.springframework.test.web.servlet.request.MockMvcRequestBuilder
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.content;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
-@UTF8MockMvcTest(controllers = {MenuGroupRestController.class})
-class MenuGroupRestControllerTest {
+@UTF8MockMvcTest(controllers = {MenuRestController.class})
+class MenuRestControllerTest {
 
   @Autowired
   private MockMvc mockMvc;
@@ -31,20 +34,20 @@ class MenuGroupRestControllerTest {
   private ObjectMapper objectMapper;
 
   @MockBean
-  private MenuGroupService menuGroupService;
+  private MenuService menuService;
 
-  @DisplayName("새로운 메뉴그룹을 저장한다.")
+  @DisplayName("새로운 메뉴를 저장한다.")
   @Test
   void createTest() throws Exception {
     //given
-    MenuGroup createdMenuGroup = new MenuGroup(1L, "추천메뉴");
-    when(menuGroupService.create(any())).thenReturn(createdMenuGroup);
-    MenuGroup createRequestBody = new MenuGroup("추천메뉴");
+    Menu createdMenu = new Menu(1L, "후라이드+후라이드", BigDecimal.valueOf(19000), 1L, Arrays.asList(new MenuProduct(1L, 1L, 2)));
+    when(menuService.create(any())).thenReturn(createdMenu);
+    Menu createRequestBody = new Menu("후라이드+후라이드", BigDecimal.valueOf(19000), 1L, Arrays.asList(new MenuProduct(1L, 2)));
     String requestBody = objectMapper.writeValueAsString(createRequestBody);
 
     //when & then
     mockMvc
-        .perform(post("/api/menu-groups")
+        .perform(post("/api/menus")
             .contentType(MediaType.APPLICATION_JSON_VALUE)
             .content(requestBody))
         .andExpect(status().isCreated());
@@ -53,13 +56,13 @@ class MenuGroupRestControllerTest {
   @DisplayName("메뉴그룹 목록을 조회한다.")
   @Test
   void listTest() throws Exception {
-    MenuGroup createdMenuGroup1 = new MenuGroup(1L, "추천메뉴");
-    MenuGroup createdMenuGroup2 = new MenuGroup(2L, "오늘의메뉴");
-    when(menuGroupService.list()).thenReturn(Arrays.asList(createdMenuGroup1, createdMenuGroup2));
+    Menu createdMenu1 = new Menu(1L, "후라이드+후라이드", BigDecimal.valueOf(19000), 1L, Arrays.asList(new MenuProduct(1L, 1L, 2)));
+    Menu createdMenu2 = new Menu(2L, "후라이드+양념", BigDecimal.valueOf(20000), 1L, Arrays.asList(new MenuProduct(2L, 2L, 1)));
+    when(menuService.list()).thenReturn(Arrays.asList(createdMenu1, createdMenu2));
     mockMvc
-        .perform(get("/api/menu-groups"))
+        .perform(get("/api/menus"))
         .andExpect(status().isOk())
-        .andExpect(content().string(containsString("추천메뉴")))
-        .andExpect(content().string(containsString("오늘의메뉴")));
+        .andExpect(content().string(containsString("후라이드+후라이드")))
+        .andExpect(content().string(containsString("후라이드+양념")));
   }
 }
